drivers: meter: ade7978: fix incompatible-pointer-types for GCC 14+
GCC 14 makes -Wincompatible-pointer-types an error. Read the register value
into a uint8_t local matching the accessor rather than a uint32_t, so the
pointer passed to the read matches its expected width.

+++ drivers/meter/ade7978/ade7978.c
@@ -304,7 +304,7 @@ int ade7978_init(struct ade7978_dev **device,
 	/* chip id read value */
 	uint32_t chip_id;
 	/* data read */
-	uint32_t data;
+	uint8_t data;
 
 	dev = (struct ade7978_dev *)no_os_calloc(1, sizeof(*dev));
 	if (!dev)
